Data Set & NoteBook Update
Did the second half of the Lepiota Mushroom species characterization for the synthetic data set.

Did a complete PCA of all dummied variables from the original dataset. Calculated and visualized explained variance ratio and cumulative explained variance ratio as well as showing all of the loadings for the different dummied variables.